requirements.txt
numpy 
lightkurve
astropy 
astroquery
pandas
pydl
python-dotenv
tensorflow